requirements.txt
matplotlib
numpy
PyYAML
requests
scikit-learn
sklearn
Flask